+++ tango/__init__.py
@@ -14,8 +14,8 @@ __all__ = [
     "LocalStepCache",
 ]
 
-from .executor import Executor, SimpleExecutor
-from .format import Format, DillFormat, DillFormatIterator, JsonFormat, JsonFormatIterator
-from .step import Step
-from .step_cache import StepCache, LocalStepCache
-from .step_graph import StepGraph, StepStub
+from tango.executor import Executor, SimpleExecutor
+from tango.format import Format, DillFormat, DillFormatIterator, JsonFormat, JsonFormatIterator
+from tango.step import Step
+from tango.step_cache import StepCache, LocalStepCache
+from tango.step_graph import StepGraph, StepStub