+++ tango/step_cache.py
@@ -1,5 +1,5 @@
 import collections
-import json
+from dataclasses import dataclass
 import logging
 import weakref
 from abc import abstractmethod
@@ -24,11 +24,12 @@ except ImportError:
         return getattr(tp, "__args__", ())
 
 
-from .common.registrable import Registrable
-from .common.util import PathOrStr
+from tango.common.from_params import FromParams
+from tango.common.registrable import Registrable
+from tango.common.util import PathOrStr
 
 if TYPE_CHECKING:
-    from .step import Step
+    from tango.step import Step
 
 logger = logging.getLogger(__name__)
 
@@ -93,8 +94,8 @@ class LocalStepCache(StepCache):
     Every cached step gets a directory under ``dir`` with that step's :attr:`~tango.step.Step.unique_id`.
     In that directory we store the results themselves in some format according to the step's
     :attr:`~tango.step.Step.FORMAT`,
-    and we also write a ``metadata.json`` file that stores some metadata. The presence of
-    ``metadata.json`` signifies that the cache entry is complete and has been written successfully.
+    and we also write a ``cache-metadata.json`` file that stores some metadata. The presence of
+    ``cache-metadata.json`` signifies that the cache entry is complete and has been written successfully.
 
     .. tip::
         Registered as :class:`StepCache` under the name "local".
@@ -148,7 +149,7 @@ class LocalStepCache(StepCache):
                 return True
             if key in self.weak_cache:
                 return True
-            metadata_file = self.path_for_step(step) / "metadata.json"
+            metadata_file = self.path_for_step(step) / "cache-metadata.json"
             return metadata_file.exists()
         else:
             return False
@@ -167,19 +168,15 @@ class LocalStepCache(StepCache):
         location = self.path_for_step(step)
         location.mkdir(parents=True, exist_ok=True)
 
-        metadata_location = location / "metadata.json"
+        metadata_location = location / "cache-metadata.json"
         if metadata_location.exists():
             raise ValueError(f"{metadata_location} already exists! Will not overwrite.")
         temp_metadata_location = metadata_location.with_suffix(".temp")
 
         try:
             step.format.write(value, location)
-            metadata = {
-                "step": step.unique_id,
-                "checksum": step.format.checksum(location),
-            }
-            with temp_metadata_location.open("wt") as f:
-                json.dump(metadata, f)
+            metadata = CacheMetadata(step=step.unique_id, checksum=step.format.checksum(location))
+            metadata.to_params().to_file(temp_metadata_location)
             self._add_to_cache(step.unique_id, value)
             temp_metadata_location.rename(metadata_location)
         except:  # noqa: E722
@@ -190,4 +187,10 @@ class LocalStepCache(StepCache):
             raise
 
     def __len__(self) -> int:
-        return sum(1 for _ in self.dir.glob("*/metadata.json"))
+        return sum(1 for _ in self.dir.glob("*/cache-metadata.json"))
+
+
+@dataclass
+class CacheMetadata(FromParams):
+    step: str
+    checksum: str